app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
boto3==1.40.4
botocore==1.40.4
certifi==2025.8.3
charset-normalizer==3.4.2
dataclasses-json==0.6.7
decorator==5.2.1
executing==2.2.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-aws==0.2.30
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langgraph-supervisor==0.0.29
langsmith==0.4.13
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.3
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
stack-data==0.6.3
tenacity==9.1.2
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0